pyproject.toml
[project]
name = "narutogame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pygame-ce>=2.5.7",
]
